RavenDB-13106 fix tests
