+++ test/SlowTests/Sharding/Cluster/ReshardingTests.cs
@@ -88,8 +88,8 @@ namespace SlowTests.Sharding.Cluster
                 var id = "orders/830-A";
                
                 var bucket = ShardHelper.GetBucket(id);
-                var location = ShardHelper.GetShardNumber(record.ShardBucketRanges, bucket);
-                var newLocation = (location + 1) % record.Shards.Length;
+                var location = ShardHelper.GetShardNumber(record.Sharding.BucketRanges, bucket);
+                var newLocation = (location + 1) % record.Sharding.Shards.Length;
                 using (var session = store.OpenAsyncSession(ShardHelper.ToShardName(store.Database, location)))
                 {
                     var order = await session.LoadAsync<Order>(id);
