+++ test/Tryouts/Program.cs
@@ -8,6 +8,7 @@ using RachisTests.DatabaseCluster;
 using Raven.Server.Utils;
 using SlowTests.Cluster;
 using SlowTests.Issues;
+using SlowTests.Sharding.Cluster;
 
 namespace Tryouts;
 
@@ -27,10 +28,10 @@ public static class Program
             try
             {
                 using (var testOutputHelper = new ConsoleTestOutputHelper())
-                using (var test = new ClusterDatabaseMaintenance(testOutputHelper))
+                using (var test = new SubscriptionsWithReshardingTests(testOutputHelper))
                 {
                     DebuggerAttachedTimeout.DisableLongTimespan = true;
-                    await test.OnlyOneNodeShouldUpdateRehab();
+                    await test.ContinueSubscriptionAfterResharding();
                 }
             }
             catch (Exception e)
